Utilizando PageView para alternar entre paginas. Cria SettingsPage.
