--- poc_amplitude/amplitude_poc/lib/pages/homepage.dart
@@ -1,62 +1,0 @@
-import 'package:amplitude_poc/app_state.dart';
-import 'package:amplitude_poc/pages/page01.dart';
-import 'package:amplitude_poc/resources/strings.dart';
-import 'package:amplitude_poc/widgets/nav_bar.dart';
-import 'package:flutter/material.dart';
-
-class HomePage extends StatefulWidget {
-  @override
-  _HomePageState createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  int _selectedIndex = 0;
-
-  void _startFlow() {
-    AppState.of(context).logEvent(
-      EventType.BUTTON_PRESSED,
-      'HomePage._onPressed',
-    );
-
-    Navigator.push(context, MaterialPageRoute(builder: (_) {
-      return PageOne();
-    }));
-  }
-
-  void _changePage(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: Text('Amplitude')),
-      body: Center(
-        child: RaisedButton.icon(
-          onPressed: _startFlow,
-          icon: Icon(Icons.adb_rounded, color: Colors.white),
-          color: Colors.pink,
-          label: Text(
-            Strings.START_FLOW,
-            style: TextStyle(
-              color: Colors.white,
-              fontSize: 16.0,
-            ),
-          ),
-        ),
-      ),
-      bottomNavigationBar: NavBar(
-        onPressed: _changePage,
-        selectedIndex: _selectedIndex,
-        items: [
-          NavBarItem(iconData: Icons.add),
-          NavBarItem(iconData: Icons.grid_on),
-          NavBarItem(iconData: Icons.timer),
-          NavBarItem(iconData: Icons.settings),
-        ],
-      ),
-    );
-  }
-}
