+++ poc_amplitude/amplitude_poc/lib/pages/settings_page.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class SettingsPage extends StatefulWidget {
+  @override
+  _SettingsPageState createState() => _SettingsPageState();
+}
+
+class _SettingsPageState extends State<SettingsPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
